examples/weather/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[openai]>=1.1.3",
    "python-dotenv>=1.2.1",
]
